+++ src/game-boy-emulator/resampler.hpp
@@ -17,8 +17,9 @@ class Resampler {
 public:
     Resampler(int format_in, int format_out, size_t samplerate_in, size_t samplerate_out,
               int channels_in, int channels_out) {
-        m_audio_stream = SDL_NewAudioStream(format_in, channels_in, samplerate_in, format_out,
-                                            channels_out, samplerate_out);
+        m_audio_stream
+            = SDL_NewAudioStream(format_in, channels_in, static_cast<int>(samplerate_in),
+                                 format_out, channels_out, static_cast<int>(samplerate_out));
         if (m_audio_stream == nullptr) {
             spdlog::error("Failed to create SDL_AudioStream: {}", SDL_GetError());
             std::exit(EXIT_FAILURE);
@@ -55,7 +56,8 @@ public:
     // Write available resampled data into the buffer.
     // The buffer size should not exceed the number of samples available.
     int get_resampled_data(std::span<Sample> buffer) {
-        return SDL_AudioStreamGet(m_audio_stream, buffer.data(), buffer.size_bytes());
+        return SDL_AudioStreamGet(m_audio_stream, buffer.data(),
+                                  static_cast<int>(buffer.size_bytes()));
     }
     std::vector<Sample> get_resampled_data() {
         const auto buffer_size = available_samples();
